Building, BuildingCatalog, Item, Item Catalog, Inventory, InventoryCatalog, Craft building, autopopulat menu,
Made framework for inventories, items and building. Adding items is done in the building, item and inventory files.
Player can now craft buildings (Still need to be tweaked).
Craftingmenu is autopopulation according to how many buildings there are in the game. Each buildings show how much the cost to craft is.