Walking updated, Autogenerating map, Updated UI
Updated walking so that it has better collision with other objects. This is because it is now using rigidbody more correctly.
The map is autogenerated when the player is walking a certain distance. The new plane that is generated is then populated with trees and deposits. The amount of trees and other resources are generated at random.
Adjusted UI ratios and changed som colors.
Other improvents to code has also been done.