New new buildings, Added interaction with buildings, UI, Structuring
Added Small stone wall, medium stone wall, medium stone gate, medium stone corner
Added UI and code for interacting with with buildings. Theres detection if the building is owned by the player, and if it is, theres som choice for what do with the building.
Updated code related to building and crafting.
Made more structured folder