Towns, Town Catalog, Citizens, Buildings, UI
Introduced the concept of town. A town is created when a town center is placed. Buildings inside the town center is a part of that town, and the citizens living in those buildings are part of the towns citizens. Buildings and NPC's can be dynamically added and removed from buildings and towns. Citizens will are assigned houses that are free, and if they are an available citizen, they can be choosen to work in a productionbuilding that is placed in the city.
Added Town Catalog, that tracks all the towns in game.
Updated citizen-script with allot of functionalitites.
Building-script have been updated accordingly.
Added preperations for NPCclickMenu, ConfirmationMenu, TownMenu

Co-Authored-By: LeonardotNEO <52072762+LeonardotNEO@users.noreply.github.com>